big cleanup + rewrite of inner buildloop
