+++ src/BuildLoop.hpp
@@ -1,0 +1,6 @@
+namespace t2
+{
+    struct ThreadState;
+    
+    void BuildLoop(ThreadState* thread_state);
+}
