+++ src/InputSignature.hpp
@@ -1,0 +1,8 @@
+namespace t2
+{
+struct ThreadState;
+struct BuildQueue;
+struct NodeState;
+
+bool CheckInputSignatureToSeeNodeNeedsExecuting(BuildQueue* queue, ThreadState* thread_state, NodeState* node);
+}
